CTCP-5346: Restructuring remaining packages.
