+++ app/controllers/TransitAccompanyingDocumentP5Controller.scala
@@ -14,17 +14,17 @@
  * limitations under the License.
  */
 
-package controllers.P5
+package controllers
 
 import controllers.actions.AuthenticateActionProvider
 import controllers.actions.VersionedAction
-import models.P5.Phase.PostTransition
-import models.P5.Phase.Transition
+import models.Phase.PostTransition
+import models.Phase.Transition
 import play.api.Logging
 import play.api.mvc.Action
 import play.api.mvc.AnyContent
 import play.api.mvc.ControllerComponents
-import services.P5.DepartureMessageP5Service
+import services.messages.DepartureMessageP5Service
 import services.pdf.TADPdfGenerator
 import uk.gov.hmrc.play.bootstrap.backend.controller.BackendController
 import utils.FileNameSanitizer
