+++ app/models/DepartureMessages.scala
@@ -14,24 +14,16 @@
  * limitations under the License.
  */
 
-package models.P5.departure
+package models
 
+import play.api.libs.json.Json
 import play.api.libs.json.Reads
-import play.api.libs.json.__
 
-import java.time.LocalDateTime
+case class DepartureMessages(messages: List[DepartureMessageMetaData]) {
 
-case class DepartureMessageMetaData(id: String, received: LocalDateTime, messageType: DepartureMessageType, path: String)
+  def find(`type`: DepartureMessageType): Option[DepartureMessageMetaData] = messages.find(_.messageType == `type`)
+}
 
-object DepartureMessageMetaData {
-
-  implicit lazy val reads: Reads[DepartureMessageMetaData] = {
-    import play.api.libs.functional.syntax._
-    (
-      (__ \ "id").read[String] and
-        (__ \ "received").read[LocalDateTime] and
-        (__ \ "type").read[DepartureMessageType] and
-        (__ \ "_links" \ "self" \ "href").read[String].map(_.replace("/customs/transits/", ""))
-    )(DepartureMessageMetaData.apply _)
-  }
+object DepartureMessages {
+  implicit val reads: Reads[DepartureMessages] = Json.reads[DepartureMessages]
 }
